Remove verifactu validations for special characters in free text fields
