+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Removed
+
+- `es-verifactu-v1`: validation to prevent forbidden characters in names. Validation only in invoice series and code.
+
 ## [v0.301.0] - 2025-10-03
 
 ### Added
