+++ addons/es/verifactu/bill_test.go
@@ -351,95 +351,19 @@ func TestInvoiceValidation(t *testing.T) {
 		require.NoError(t, inv.Calculate())
 		require.NoError(t, inv.Validate())
 	})
-}
 
-func TestForbiddenCharactersValidation(t *testing.T) {
-	t.Run("supplier name with forbidden char =", func(t *testing.T) {
+	t.Run("code with forbidden characters", func(t *testing.T) {
 		inv := testInvoiceStandard(t)
-		inv.Supplier.Name = "Test Supplier ="
+		inv.Code = "123<"
 		require.NoError(t, inv.Calculate())
-		err := inv.Validate()
-		require.ErrorContains(t, err, "contains forbidden character: =")
+		assert.ErrorContains(t, inv.Validate(), "code: contains forbidden character: <")
 	})
 
-	t.Run("supplier name with forbidden char '", func(t *testing.T) {
+	t.Run("series with forbidden characters", func(t *testing.T) {
 		inv := testInvoiceStandard(t)
-		inv.Supplier.Name = "Test Supplier '"
+		inv.Series = "SAMPLE="
 		require.NoError(t, inv.Calculate())
-		err := inv.Validate()
-		require.ErrorContains(t, err, "contains forbidden character: '")
-	})
-
-	t.Run("customer name with forbidden char >", func(t *testing.T) {
-		inv := testInvoiceStandard(t)
-		inv.Customer.Name = "Test Customer >"
-		require.NoError(t, inv.Calculate())
-		err := inv.Validate()
-		require.ErrorContains(t, err, "contains forbidden character: >")
-	})
-
-	t.Run("ordering issuer name with forbidden char <", func(t *testing.T) {
-		inv := testInvoiceStandard(t)
-		inv.Ordering = &bill.Ordering{
-			Issuer: &org.Party{
-				Name: "Test Issuer <",
-				TaxID: &tax.Identity{
-					Country: "ES",
-					Code:    "B98602642",
-				},
-			},
-		}
-		require.NoError(t, inv.Calculate())
-		err := inv.Validate()
-		require.ErrorContains(t, err, "contains forbidden character: <")
-	})
-
-	t.Run("note text with forbidden char \"", func(t *testing.T) {
-		inv := testInvoiceStandard(t)
-		inv.Notes = []*org.Note{
-			{
-				Key:  org.NoteKeyGeneral,
-				Text: "Test note \"",
-			},
-		}
-		require.NoError(t, inv.Calculate())
-		err := inv.Validate()
-		require.ErrorContains(t, err, "contains forbidden character: \"")
-	})
-
-	t.Run("valid names without forbidden characters pass validation", func(t *testing.T) {
-		inv := testInvoiceStandard(t)
-		inv.Supplier.Name = "Valid Supplier Name 123"
-		inv.Customer.Name = "Valid Customer Name 456"
-		inv.Ordering = &bill.Ordering{
-			Issuer: &org.Party{
-				Name: "Valid Issuer Name 789",
-				TaxID: &tax.Identity{
-					Country: "ES",
-					Code:    "B98602642",
-				},
-			},
-		}
-		inv.Notes = []*org.Note{
-			{
-				Key:  org.NoteKeyGeneral,
-				Text: "Valid note text with numbers 123 and symbols !@#$%^&*()_+-{}[]|\\:;.,?/~`",
-			},
-		}
-		require.NoError(t, inv.Calculate())
-		require.NoError(t, inv.Validate())
-	})
-
-	t.Run("non-general note with forbidden characters is ignored", func(t *testing.T) {
-		inv := testInvoiceStandard(t)
-		inv.Notes = []*org.Note{
-			{
-				Key:  org.NoteKeyLoading,
-				Text: "Loading note with forbidden chars: < > \" ' =",
-			},
-		}
-		require.NoError(t, inv.Calculate())
-		require.NoError(t, inv.Validate())
+		assert.ErrorContains(t, inv.Validate(), "series: contains forbidden character: <")
 	})
 }
