+++ addons/es/verifactu/verifactu.go
@@ -5,7 +5,6 @@ import (
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/i18n"
-	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/pkg/here"
 	"github.com/invopop/gobl/tax"
 )
@@ -79,8 +78,6 @@ func validate(doc any) error {
 		return validateInvoice(obj)
 	case *tax.Combo:
 		return validateTaxCombo(obj)
-	case *org.Party:
-		return validateParty(obj)
 	}
 	return nil
 }
