+++ data/regimes/de.json
@@ -46,7 +46,7 @@
     {
       "code": "VAT",
       "name": {
-        "de": "MwSt",
+        "de": "USt",
         "en": "VAT"
       },
       "title": {
